+++ parser/CHANGELOG.md
@@ -10,7 +10,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 
 - Add support for enum and struct fallbacks.
-- Add error type `DuplicateEnumFallbackName`.
+- Add error types `DuplicateEnumFallbackName` and `DuplicateStructFallbackName`.
 - Add warning type `NonCamelCaseEnumFallback`.
 
 ### Fixed